Struct rhymer::Document

pub struct Document { /* fields omitted */ }

A BSON document represented as an associative HashMap with insertion ordering.

Implementations

impl Document

pub fn new() -> Document

Creates a new empty Document.

pub fn iter(&self) -> DocumentIterator<'_>

Gets an iterator over the entries of the map.

pub fn clear(&mut self)

Clears the document, removing all values.

pub fn get(&self, key: &str) -> Option<&Bson>

Returns a reference to the Bson corresponding to the key.

pub fn get_mut(&mut self, key: &str) -> Option<&mut Bson>

Gets a mutable reference to the Bson corresponding to the key

pub fn get_f64(&self, key: &str) -> Result<f64, ValueAccessError>

Get a floating point value for this key if it exists and has the correct type.

pub fn get_f64_mut(&mut self, key: &str) -> Result<&mut f64, ValueAccessError>

Get a mutable reference to a floating point value for this key if it exists and has the correct type.

pub fn get_str(&self, key: &str) -> Result<&str, ValueAccessError>

Get a string slice this key if it exists and has the correct type.

pub fn get_str_mut(&mut self, key: &str) -> Result<&mut str, ValueAccessError>

Get a mutable string slice this key if it exists and has the correct type.

pub fn get_array(&self, key: &str) -> Result<&Vec<Bson>, ValueAccessError>

Get a reference to an array for this key if it exists and has the correct type.

pub fn get_array_mut(
&mut self,
key: &str
) -> Result<&mut Vec<Bson>, ValueAccessError>

Get a mutable reference to an array for this key if it exists and has the correct type.

pub fn get_document(&self, key: &str) -> Result<&Document, ValueAccessError>

Get a reference to a document for this key if it exists and has the correct type.

pub fn get_document_mut(
&mut self,
key: &str
) -> Result<&mut Document, ValueAccessError>

Get a mutable reference to a document for this key if it exists and has the correct type.

pub fn get_bool(&self, key: &str) -> Result<bool, ValueAccessError>

Get a bool value for this key if it exists and has the correct type.

pub fn get_bool_mut(&mut self, key: &str) -> Result<&mut bool, ValueAccessError>

Get a mutable reference to a bool value for this key if it exists and has the correct type.

pub fn is_null(&self, key: &str) -> bool

Returns wether this key has a null value

pub fn get_i32(&self, key: &str) -> Result<i32, ValueAccessError>

Get an i32 value for this key if it exists and has the correct type.

pub fn get_i32_mut(&mut self, key: &str) -> Result<&mut i32, ValueAccessError>

Get a mutable reference to an i32 value for this key if it exists and has the correct type.

pub fn get_i64(&self, key: &str) -> Result<i64, ValueAccessError>

Get an i64 value for this key if it exists and has the correct type.

pub fn get_i64_mut(&mut self, key: &str) -> Result<&mut i64, ValueAccessError>

Get a mutable reference to an i64 value for this key if it exists and has the correct type.

pub fn get_timestamp(&self, key: &str) -> Result<Timestamp, ValueAccessError>

Get a time stamp value for this key if it exists and has the correct type.

pub fn get_timestamp_mut(
&mut self,
key: &str
) -> Result<&mut Timestamp, ValueAccessError>

Get a mutable reference to a time stamp value for this key if it exists and has the correct type.

pub fn get_binary_generic(
&self,
key: &str
) -> Result<&Vec<u8>, ValueAccessError>

Get a reference to a generic binary value for this key if it exists and has the correct type.

pub fn get_binary_generic_mut(
&mut self,
key: &str
) -> Result<&mut Vec<u8>, ValueAccessError>

Get a mutable reference generic binary value for this key if it exists and has the correct type.

pub fn get_object_id(&self, key: &str) -> Result<&ObjectId, ValueAccessError>

Get a reference to an object id value for this key if it exists and has the correct type.

pub fn get_object_id_mut(
&mut self,
key: &str
) -> Result<&mut ObjectId, ValueAccessError>

Get a mutable reference to an object id value for this key if it exists and has the correct type.

pub fn get_datetime(
&self,
key: &str
) -> Result<&DateTime<Utc>, ValueAccessError>

Get a reference to a UTC datetime value for this key if it exists and has the correct type.

pub fn get_datetime_mut(
&mut self,
key: &str
) -> Result<&mut DateTime<Utc>, ValueAccessError>

Get a mutable reference to a UTC datetime value for this key if it exists and has the correct type.

pub fn contains_key(&self, key: &str) -> bool

Returns true if the map contains a value for the specified key.

pub fn keys(&'a self) -> Keys<'a>

Gets a collection of all keys in the document.

pub fn values(&'a self) -> Values<'a>

Gets a collection of all values in the document.

pub fn len(&self) -> usize

Returns the number of elements in the document.

pub fn is_empty(&self) -> bool

Returns true if the document contains no elements

pub fn insert<KT, BT>(&mut self, key: KT, val: BT) -> Option<Bson> where
BT: Into<Bson>,
KT: Into<String>,

Sets the value of the entry with the OccupiedEntry's key, and returns the entry's old value. Accepts any type that can be converted into Bson.

pub fn remove(&mut self, key: &str) -> Option<Bson>

Takes the value of the entry out of the document, and returns it.

pub fn entry(&mut self, k: String) -> Entry<'_>

pub fn to_writer<W>(&self, writer: &mut W) -> Result<(), Error> where
W: Write + ?Sized,

Attempts to serialize the Document into a byte stream.

pub fn from_reader<R>(reader: &mut R) -> Result<Document, Error> where
R: Read + ?Sized,

Attempts to deserialize a Document from a byte stream.